requirements.txt
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
pytest==8.3.3
